Add a gui directory in core and move widgets there.
