+++ src/core/gui/widgets/BoxLayout.cpp
