+++ src/core/gui/widgets/ClickableWidget.cpp
