+++ src/core/gui/widgets/Group.cpp
