+++ src/core/gui/widgets/HorizontalLayout.cpp
