+++ src/core/gui/widgets/Label.cpp
