+++ src/core/gui/widgets/Panel.cpp
