+++ src/core/gui/widgets/Panel.h
