+++ src/core/gui/widgets/VerticalLayout.cpp
