+++ src/core/gui/widgets/VerticalLayout.h
