+++ src/core/gui/widgets/WidgetType.cpp
