+++ src/core/gui/widgets/WidgetType.h
