+++ tests/core/widgets/ButtonTests.cpp
@@ -1,6 +1,7 @@
 #include <catch2/catch.hpp>
 
-#include "core/widgets/Button.h"
+#include "core/gui/widgets/Button.h"
+
 #include "../TestsUtils.h"
 
 TEST_CASE("[Button]")
