+++ tests/core/widgets/PanelTests.cpp
@@ -1,9 +1,11 @@
 #include <catch2/catch.hpp>
 
-#include "core/widgets/Panel.h"
-#include "core/widgets/VerticalLayout.h"
-#include "core/widgets/Button.h"
+#include "core/gui/widgets/Panel.h"
+#include "core/gui/widgets/VerticalLayout.h"
+#include "core/gui/widgets/Button.h"
+
 #include "../TestsUtils.h"
+
 TEST_CASE("[Panel]")
 {
